PCL2 Scriptable Object configs
