+++ com.unity.robotics.message-visualizations/Runtime/DefaultVisualizers/Sensor/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d43d593c8032438cbf9587f51523ae2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
