+++ com.unity.robotics.message-visualizations/Runtime/DefaultVisualizers/Sensor/Editor/DefaultVisualizerPointCloud2Editor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afb6421e5f624409d8d7530743198fbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
