+++ com.unity.robotics.message-visualizations/Runtime/DefaultVisualizers/Sensor/ScriptableObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 446495620753543beaf7487d9451c1bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
