+++ com.unity.robotics.message-visualizations/Runtime/DefaultVisualizers/Sensor/ScriptableObjects/PointCloud2VisualizerSettings.asset
@@ -1,0 +1,41 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9869b969f021440789a37afecf3ffca4, type: 3}
+  m_Name: PointCloud2VisualizerSettings
+  m_EditorClassIdentifier: 
+  colorMode: 0
+  m_XChannel: x
+  m_YChannel: y
+  m_ZChannel: z
+  m_RgbChannel: ring
+  m_RChannel: x
+  m_GChannel: y
+  m_BChannel: z
+  m_SizeChannel: intensity
+  m_RgbRange:
+  - 0
+  - 31
+  m_RRange:
+  - -100
+  - 100
+  m_GRange:
+  - -100
+  - 100
+  m_BRange:
+  - -100
+  - 100
+  m_SizeRange:
+  - 0
+  - 100
+  m_Size: 0.01
+  m_UseRgbChannel: 0
+  m_UseSizeChannel: 0
