+++ com.unity.robotics.message-visualizations/Runtime/DefaultVisualizers/Sensor/ScriptableObjects/PointCloud2VisualizerSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1d2df8bcd83d4410b03e8a6ddd5eac4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
